Added InformationAndPriceCard component and use in mobile and desktop
